Read metadata from metadata.csv
Successfully reading metadata from a CSV file.
Names/quantity of river basins is not hardcoded; the python script will automatically find unique values for the Basin column in metadata.csv and create an index.
Colours (as used to colour-code the site points on the map, and the river name legend at the top-right) are assigned automatically from a list of 13. Max number of basins is therefore 13. (I picked the colours from the supported named colours in folium, taking out any that were too generic - e.g. white, grey, beige, black - and reordering them).
Have named the "Site ID" column in metadata.csv to "SiteID" (no space) because the space was making it impossible to use the value as needed.
The script now sorts each site's river data after it reads it from the CSV. I noticed the rows weren't in order of date/time and I suspect this was causing some graphing errors.