Separate rainfall and height axes
Rainfall bumped out to a third axis. It's right-aligned because I couldn't avoid the axis label being cut off when I put it on the left (idk why this happened). I did this to make sure the flow and height (when autoscaled) could be aligned, without rainfall scaling messing it up.
Legend location is also dynamic. It defaults to top-right for all available variables, but if the graphed data would cover the legend, it moves elsewhere.
Keeping this branched off for now because I'm not sure if this is better than the combined axis, or just different. If the axes aren't autoscaled, this is sort of pointless.